# openapi.yaml
openapi: 3.0.3
info:
  title: AgriNex - Agricultural IoT Data Management Platform
  description: |
    AgriNex是一个农业物联网数据管理平台，采用三层架构：设备(Device) → 传感器(Sensor) → 读数(Reading)。
    
    核心功能：
    - 支持多类型传感器：数值型（温度、湿度、光照）、多媒体型（图片、视频）
    - MinIO对象存储集成：支持图片/视频文件的对象存储和本地备份
    - JWT认证系统：完整的用户注册登录系统
    - 告警管理：实时异常数据监控和告警
    - 预测分析：基于历史数据的预测功能
    - MCP服务：模型上下文协议集成
    
    API架构：
    - 主服务：平台信息、健康检查、API文档
    - 设备管理：完整的CRUD操作和统计分析
    - 传感器管理：传感器CRUD、读数管理、统计信息
    - 告警系统：告警列表、解决、统计
    - 预测服务：手动触发预测、获取预测结果
    - 用户认证：注册、登录、token刷新、用户信息
    - MCP服务：模型上下文协议相关接口
  version: 1.0.0
  contact:
    name: AgriNex Development Team
    email: dev@agrinex.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.agrinex.com
    description: Production server
...
security:
  - bearerAuth: []
paths:
...
  /api/devices/{device_id}/sensors:
    get:
      tags: [Devices]
      summary: 获取设备传感器
      description: 获取设备下的所有传感器
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 传感器列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Sensor'

    post:
      tags: [Devices]
      summary: 创建传感器
      description: 在设备下创建新传感器
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type:
                  type: string
                name:
                  type: string
                unit:
                  type: string
                status:
                  type: string
                  default: active
      responses:
        '201':
          description: 传感器创建成功
...
  /api/devices/{device_id}/sensors/{sensor_id}/readings:
    post:
      tags: [Devices]
      summary: 创建传感器读数
      description: 为指定传感器创建新的读数记录
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: integer
        - name: sensor_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                data_type:
                  type: string
                  enum: [numeric, image, video]
                value:
                  type: number
                unit:
                  type: string
                file:
                  type: string
                  format: binary
                metadata:
                  type: string
      responses:
        '201':
          description: 读数创建成功
...
  /api/sensors/{sensor_id}:
    get:
      tags: [Sensors]
      summary: 获取传感器详情
      description: 获取指定传感器的详细信息
      parameters:
        - name: sensor_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 传感器详情
  /api/sensors/{sensor_id}/readings:
    get:
      tags: [Sensors]
      summary: 获取传感器读数
      description: 获取指定传感器的读数列表
      parameters:
        - name: sensor_id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
        - name: start_time
          in: query
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: 读数列表

    post:
      tags: [Sensors]
      summary: 创建传感器读数
      description: 为传感器创建新的数值型读数
      parameters:
        - name: sensor_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value:
                  type: number
                unit:
                  type: string
                metadata:
                  type: object
      responses:
        '201':
          description: 读数创建成功
  /api/sensors/{sensor_id}/readings/latest:
    get:
      tags: [Sensors]
      summary: 获取最新读数
      description: 获取传感器的最新读数
      parameters:
        - name: sensor_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 最新读数
  /api/sensors/{sensor_id}/statistics:
    get:
      tags: [Sensors]
      summary: 传感器统计信息
      description: 获取传感器的统计分析信息
      parameters:
        - name: sensor_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 统计信息

  # Alarm endpoints
...
  /api/alarms/sensors/{sensor_id}:
    get:
      tags: [Alarms]
      summary: 获取传感器告警
      description: 获取指定传感器的告警列表
      parameters:
        - name: sensor_id
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, resolved, all]
      responses:
        '200':
          description: 传感器告警列表
...
  /api/forecast/sensors/{sensor_id}:
    post:
      tags: [Predictions]
      summary: 触发预测
      description: 手动触发传感器数据预测
      parameters:
        - name: sensor_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '202':
          description: 预测任务已提交
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  sensor_id:
                    type: integer
                  status:
                    type: string

    get:
      tags: [Predictions]
      summary: 获取预测结果
      description: 获取传感器的预测结果列表
      parameters:
        - name: sensor_id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: 预测结果列表
...
  /api/mcp/devices/{device_id}/sensors:
    get:
      tags: [MCP]
      summary: MCP获取设备传感器
      description: 通过MCP服务获取设备传感器
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 传感器列表

    post:
      tags: [MCP]
      summary: MCP添加传感器
      description: 通过MCP服务添加传感器
      parameters:
        - name: device_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type]
              properties:
                type:
                  type: string
                name:
                  type: string
                unit:
                  type: string
                status:
                  type: string
      responses:
        '201':
          description: 传感器添加成功
  /api/mcp/sensors/{sensor_id}/readings:
    get:
      tags: [MCP]
      summary: MCP获取传感器读数
      description: 通过MCP服务获取传感器读数
      parameters:
        - name: sensor_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 读数列表

    post:
      tags: [MCP]
      summary: MCP添加读数
      description: 通过MCP服务添加传感器读数
      parameters:
        - name: sensor_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                timestamp:
                  type: string
                  format: date-time
                value:
                  type: number
      responses:
        '201':
          description: 读数添加成功
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 使用JWT Bearer token进行认证
  schemas:
...
    Sensor:
      type: object
      properties:
        id:
          type: integer
        device_id:
          type: integer
        type:
          type: string
        name:
          type: string
        unit:
          type: string
        status:
          type: string
          enum: [active, inactive, error]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...